backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.12.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.3
torch==2.1.2
transformers==4.37.0
fashion-clip==0.2.2

